HealthBarUpdate
-Updated visual look and functionality
-Reworked playerHealth script to be more simple, also made new for healtBar
-Cleaned up a bit, made some new folders
