+++ Shadow Commandos/Assets/Materials/Fonts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5a71d5cb295882745afe493282a88fea
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
