+++ Shadow Commandos/Assets/Materials/Fonts/Sortelo Italic SDF.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0bf1c7dd33330a54eb7ee574b8b891ff
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
