+++ Shadow Commandos/Assets/Materials/Fonts/Sortelo Regular SDF.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c5c66a851aacbdf42b48d7b40f728f59
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
