+++ Shadow Commandos/Assets/Materials/Fonts/Sortelo Regular.otf.meta
@@ -1,0 +1,21 @@
+fileFormatVersion: 2
+guid: 0ad42a3bab8926a4a9e8e8dc7fad095d
+TrueTypeFontImporter:
+  externalObjects: {}
+  serializedVersion: 4
+  fontSize: 16
+  forceTextureCase: -2
+  characterSpacing: 0
+  characterPadding: 1
+  includeFontData: 1
+  fontNames:
+  - Sortelo
+  fallbackFontReferences: []
+  customCharacters: 
+  fontRenderingMode: 0
+  ascentCalculationMode: 1
+  useLegacyBoundsCalculation: 0
+  shouldRoundAdvanceValue: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
