+++ Shadow Commandos/Assets/Materials/Health.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8e3c62e02580dbc4e921d77dfcc2f35a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
