+++ Shadow Commandos/Assets/Materials/Health/MedPack_Red.mat.meta
