+++ Shadow Commandos/Assets/Materials/Health/MedPack_White.mat
