+++ Shadow Commandos/Assets/Scripts/Health&Damage/PlayerHealth.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: c7487574be1b3254e92b4c4e17061125
+guid: 7b6ad4c98a69ba644ac96d14e0f2d290
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
